Create netsalary.java
write a program to calculate the net payable salary using Inheritance, Constructors and this pointer. The program will take basic salary as an 
input fromthe user and calculate the total salary along with the following components: 
HRA, which is 50% of the basic salary 
Special allowance, which is 75% of the basic salary 
PF deduction. which is 12% of the basic salary 
Ater calculating the preceding salary components, the net payable salary should be displayed. Help David to 
perform the preceding requirement.